runfix: Implement proper Modal behaviour
